+++ dyndnsc/notifications/growl/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .growl import Growl
-
-
-def create_notify_handler():
-    return Growl().create_notify_handler()