+++ dyndnsc/notifications/osxnotificationcenter/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .osxnotificationcenter import OSXNotification
-
-
-def create_notify_handler():
-    return OSXNotification().create_notify_handler()